fix(frontend): trim Monaco graph for guarded shell build
